🌱 Split workloads.test.ts into 3 focused test files
Split 1840-line workloads.test.ts into:
- workloads.core.test.ts (~920 lines): usePods, useAllPods, usePodIssues, useDeployments, useJobs, etc.
- workloads.extended.test.ts (~830 lines): extended tests for all workload hooks
- workloads.branches.test.ts (~400 lines): additional branch coverage, subscribeWorkloadsCache

All 130 tests pass. Zero new lint warnings. Part of #10439.
